server/requirements.txt
fastapi
cohere
uvicorn[standard]
ffmpeg
moviepy
pytube
deepgram-sdk
python-multipart
python-dotenv